requirements.txt
future==0.18.2
python-whois==0.8.0
tomlkit==0.11.6
